added filter for all properties - Adithya S K

+++ src/main/java/com/example/coursurf/controllers/CourseController.java
@@ -42,7 +42,7 @@ public class CourseController {
     public List<Course> filterCourses(
             @RequestParam(required = false) String searchQuery,
             @RequestParam(required = false) String provider,
-            @RequestParam(required = false) Integer rating,
+            @RequestParam(required = false) Float rating,
             @RequestParam(required = false, defaultValue = "10") int limit) {
 
         return courseService.filterCourses(searchQuery, provider, rating, limit);
